Update manifest: launchMode singleTop
use launchMode:singleTop and alwaysRetainState to attempt to fix
home button bug: on some versions of Android, when you exit the app via
the home button or by switching windows and then return to it, the app
restarts.

This seems to improve matters on Android < 4.4, but on 4.4 we
still see the issue.